Externalized getValue && onTransitionFactory
